style: rename partials dir to components
