+++ src/views/_proxied-dot-io/product-docs-landing/components/featured-card/index.tsx
